+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fixes to Fallout 3/New Vegas rendering: depth buffer flags are now checked only on BSShaderNoLightingProperty, added support for Lighting30ShaderProperty and TallGrassShaderProperty.
 * Fixed the light direction being reset on changes to the render settings.
 * Fixed loading Fallout 76 and Starfield cube maps with legacy DDS header.
+* Fixed restoring the background color after silhouette mode.
 * Minor optimizations: faster cube map filtering, and use of SIMD data types in 'noavx' GCC and Clang builds.
 
 #### NifSkope-2.0.dev9-20250130